+++ src/main/scala/org/mccandless/advent/intcode/Oracle.scala
@@ -2,5 +2,5 @@ package org.mccandless.advent.intcode
 
 object Oracle {
 
-  def input(): Int = 5
+  def input(): Long = 5
 }
